# [WIP] Big speedup from considering dependency possibilities in groups