--- a/lib/molinillo/modules/ui.rb
+++ b/lib/molinillo/modules/ui.rb
@@ -48,7 +48,7 @@ module Molinillo
       if debug?
         debug_info = yield
         debug_info = debug_info.inspect unless debug_info.is_a?(String)
-        output.puts debug_info.split("\n").map { |s| ' ' * depth + s }
+        output.puts debug_info.split("\n").map { |s| ":#{depth.to_s.rjust 4}: #{s}" }
       end
     end