--- a/spec/spec_helper/index.rb
+++ b/spec/spec_helper/index.rb
@@ -193,7 +193,7 @@ module Molinillo
     BundlerIndex,
     ReverseBundlerIndex,
     BundlerSingleAllNoPenaltyIndex,
-    RandomSortIndex,
+    # RandomSortIndex,
     CocoaPodsIndex,
     BerkshelfIndex,
   ].freeze