… ARCHITECTURE.md
… CHANGELOG.md
… Gemfile.lock
… spec/spec_helper/naive_resolver.rb