# Summary
⚠️  **Totally experimental** ⚠️ 

This branch is experimenting with considering multiple dependency possibilities
at once, in order to speed up resolution. If it works, it could be a lot faster
than the current setup.

# Theory
Molinillo employs a Backtracing algorithm to attempt every possible set of
dependencies. Because this is a very large set, it finishes early if/when it
finds a resolution, and uses adds some optimising behaviour (swapping) to try
to speed up resolution at the expense of always finding a solution.

In general, the way to speed up a Backtracing algorithm is to exploit knowledge
about the solution set or constraints. The knowledge being exploited to try to
speed things up here is that **dependencies usually don't change their
sub-dependency requirements when they update**. As such, we can treat several
versions of the same dependency as functionally equivalent when resolving the
dependency tree.

# Early benchmarks

| Spec | Old number of steps | New number of steps |
-------|----------------------|------------------------|
| can resolve when two specs have the same dependencies | 1390 | 207 |
| can resolve when two specs have the same dependencies and swapping happens | 1082 | 110 |


# TODO
- [x] Extremely rough grouping code
- [x] Remove swapping code (should be covered by grouping anyway)
- [ ] Explore adding some sorting within spec groupings. Might need changes to
     integrating apps (CocoaPods, Bundler, etc.). Might not.
- [ ] Benchmark on real Gemfiles
- [ ] Clean everything up to the point where it might ever be accepted as a PR