@greysteil 2017-07-15
@segiddins - FYI. This is my final attempt to speed things up. Did you consider this approach already? Should remove the need for any of the swapping code - if we treat dependencies with the same sub-dependencies as groups, that then get refined down as resolution happens, swapping will be implicit and automatic.