@greysteil 2017-07-15
Yep - it does some aggressive filtering based on existing activated specs, which doesn't fit with this approach. Hoping we can change that...