@segiddins 2017-07-15
> Yep - it does some aggressive filtering based on existing activated specs, which doesn't fit with this approach. Hoping we can change that...

We can see what we can do, but CocoaPods needs that because of its notion of "subspecs", where each spec / subspec with the same "root name" must be at the same version