@greysteil 2017-07-16
That should be workable - [this update](https://github.com/CocoaPods/Molinillo/pull/69/commits/90f4233e289acb6fa33d82dbfbd48bfdd5e87b54), along with a tweak to `Resolution#resolve_activated_specs` should give us that behaviour in CocoaPods, whilst keeping the possibility sets approach here.