@greysteil 2017-07-17
This is getting close now, and I'd love some feedback on it. In particular, it would be great to get some guidance on the following:
1. This is a pretty large refactor now, touching some obviously critical code. It's hard to see an easy way to split it up (aside from splitting out https://github.com/CocoaPods/Molinillo/pull/67, which is already a separate PR). Is there are way to make merging this possible?
2. Does the updated approach, as detailed in the changes to `ARCHITECTURE.md`, make sense to everyone?
3. What performance tradeoffs are acceptable? I haven't benchmarked comprehensively, or tried to optimise the code for performance, but at the moment I'm seeing resolution times **increase** by 100% for simple resolutions (where the number of steps with this approach is the same as with the previous approach) and **decrease** by 66% for gnarly cases (like the "can resolve when two specs have the same dependencies" spec). Most real-life Gemfiles I've seen appear to be of the simple variety, so this change might slow things down slightly for them (although we're only talking about 1 second or so).
4. Does anyone have any currently unresolvable Gemfiles I can try this branch against? It fixes https://github.com/CocoaPods/Resolver-Integration-Specs/pull/11, and I think will fix all other erroneous `VersionConflict` errors - switching from the "swapping" strategy to this one makes it much, much easier to be certain we're rigorously checking all possibilities.

@segiddins - particularly interested in your thoughts.