@segiddins 2017-07-17
@greysteil <3 for all this work. Sorry I've been unresponsive, I was traveling the past week. I'd love to get you into the Bundler slack to have a place that's better for more in-depth discussion -- if you're interested, shoot me an email at segiddins@segiddins.me with your email and I'll invite you.