@segiddins 2017-07-17
> 1. This is a pretty large refactor now, touching some obviously critical code. It's hard to see an easy way to split it up (aside from splitting out #67, which is already a separate PR). Is there a way to make merging this possible?

Don't worry about the size of the PR -- I'm happy to work with you to get it mergeable (as long as we keep tests passing).

> 3. What performance tradeoffs are acceptable? 

I'll dig in and see what the actual differences are, and maybe with a second set of eyes we can work to reduce the cost in "simple" cases. If the improvements actually are that large on more complex cases _and_ we improve correctness, I'm willing to tolerate a small slowdown for the most simple cases.

> 4. Are there more examples of erroneously unresolvable Gemfiles I can try this branch against?

I think there are some in the Bundler repo:

- https://github.com/bundler/bundler/issues/5644 (this one may be a separate bug regarding platforms, I'm not sure)
- https://github.com/bundler/bundler/issues/5642 (which I could not reproduce to begin with)
- https://github.com/bundler/bundler/issues/5633 (which made me tear my hair out)
- https://github.com/bundler/bundler/issues/5830 (insufficient info to reproduce)

I also test all incoming PRs / releases against Bundler, RubyGems, and CocoaPods to try and catch regressions there