@dnkoutso 2017-07-17
We have a relatively large CocoaPods project where resolution takes about half the time of the total 45 second time for `pod install` to complete.

I will clone your fork and try it out for you and also measure it.