@greysteil 2017-07-17
Sweet. Working against some Gemfiles I've found this branch currently has an issue with multi-platform projects, I believe due to the list of activated platforms changing during resolution. Sure it can fixed, but FYI.

Also, since Bundler patches `Molinillo::VersionConflict#error`, you'll need to change the `t << %( was resolved to #{spec.version}, which)` line to `t << %( was resolved to #{spec.possibilities.last.version}, which)` in `lib/bundler/resolver.rb`. I guess that's a breaking change in this PR at the moment...