@segiddins 2017-07-17
> I guess that's a breaking change in this PR at the moment...

Breaking changes are fine, we're still pre-1.0 and a major rewrite like this would warrant a `0.6` anyways